+++ CHANGELOG.md
@@ -7,6 +7,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unpublished](https://github.com/metreeca/blue/compare/v0.9.0...HEAD)
 
+### Added
+
+- Accept plain string and string array shorthands for language-tagged Local/Locals values, normalising them to the
+  `und` (Undetermined) language tag
+
 ## [0.9.0](https://github.com/metreeca/blue/releases/tag/v0.9.0)
 
 Initial release.